move spot market to its own file
