+++ src/state/mod.rs
@@ -1,4 +1,5 @@
 pub mod asset;
 pub mod order;
 pub mod spot_clearinghouse;
+pub mod spot_market;
 pub mod state;
